extract horizon comparison parser
